+++ src/config/config.js
@@ -172,6 +172,12 @@ const config = convict({
 			at the API calls generated in his we browser (for example, using the web developer tools) 
 			he is able to see passwords in plain text in the body of the API requests. */    
       default: 2048
+    },
+    ssh_enable: {
+      doc: 'Flag to be able to communicate via ssh',
+      format: Boolean,
+      env: 'SESSION_SSH_COMMUNICATION',
+      default: true
     }
 	},