+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] -
+### Added
+- Configuration for limit the amount of fwclouds that we can create and, into each fwcloud, the amount of firewalls, clusters and nodes within a cluster.
 
 ### Added
 - API call for gather FWC-Agent information.