+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] -
+## [1.6.1] - 2022-10-06
 ### Added
-- Configuration for limit the amount of fwclouds that we can create and, into each fwcloud, the amount of firewalls, clusters and nodes within a cluster.
-- Configuration that allows disabling ssh communication for security reasons.
+- Configuration that allows disabling communication with firewalls by means of SSH protocol.
+- Configuration options for limit the amount of FWClouds that can be created and, into each fwcloud, the amount of firewalls, clusters and nodes within a cluster.
+
 
 ## [1.6.0] - 2022-09-22
 ### Added